Meant to be uint8_t on the laser

+++ project1/common/packet.cpp
@@ -2,7 +2,7 @@
 
 #include <stdint.h>
 
-Packet create_packet(uint16_t speedX, uint16_t speedY, uint16_t laserOn) {
+Packet create_packet(uint16_t speedX, uint16_t speedY, uint8_t laserOn) {
     Packet p = {
         .speedX = speedX,
         .speedY = speedY,
